requirements.txt
anyio==4.10.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tinydb==4.8.0
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.2
urllib3==1.26.6
